SSR; SEO; Vercel tweaks

+++ src/app/components/playlist/playlist.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectorRef, Component, OnInit, TemplateRef, ViewChild } from '@angular/core';
+import { ChangeDetectorRef, Component, Input, OnInit, TemplateRef } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { PlaylistService } from '../../services/playlist.service';
 import { AlbumService } from '../../services/album.service';
@@ -10,11 +10,13 @@ import { MixService } from '../../services/mix.service';
 import { AlbumSummary } from '../../models/Album/AlbumSummary.model';
 import { PlaylistSummary } from '../../models/Playlist/PlaylistSummary.mode';
 import { AudioService } from '../../services/audio.service';
-import { interval, map, switchMap, takeWhile } from 'rxjs';
+import { interval, switchMap, takeWhile } from 'rxjs';
 import { JobService } from '../../services/job.service';
 import { Roles } from '../../enums/roles.enum';
 import { SongSource } from '../../services/song-sources/song-source.interface';
 import { AlbumSongSource, PlaylistSongSource } from '../../services/song-sources/album-song-source';
+import { SafeBrowserService } from '../../services/safe-storage.service';
+import { SeoService } from '../../services/seo.service';
 
 @Component({
     selector: 'app-playlist',
@@ -22,9 +24,9 @@ import { AlbumSongSource, PlaylistSongSource } from '../../services/song-sources
     styleUrls: ['./playlist.component.css']
 })
 export class PlaylistComponent implements OnInit {
-    playlist: Playlist | Album | null = null;
     type = "";
     isDownloading: boolean = false;
+    @Input() playlist: Playlist | Album | null = null;
 
     constructor(
         private route: ActivatedRoute,
@@ -36,7 +38,9 @@ export class PlaylistComponent implements OnInit {
         private mixService: MixService,
         private audioService: AudioService,
         private jobService: JobService,
-        private cdRef: ChangeDetectorRef
+        private cdRef: ChangeDetectorRef,
+        private safeBrowserService: SafeBrowserService,
+        private seoService: SeoService
     ) { }
 
     roles = Roles;
@@ -44,40 +48,56 @@ export class PlaylistComponent implements OnInit {
 
 
     ngOnInit(): void {
-        this.route.paramMap.subscribe(params => {
-            const guid = params.get('guid');
-            this.type = this.route.snapshot.data['type'];
+        this.type = this.route.snapshot.data['type'];
+        this.playlist = this.route.snapshot.data['playlist'];
 
-            if (guid) {
-                this.songSource = this.type === 'playlist'
-                    ? new PlaylistSongSource(this.playlistService, guid)
-                    : new AlbumSongSource(this.albumService, guid);
+        if (this.playlist) {
+            this.initSeo();
 
-                this.fetchData(guid);
-            }
-        });
+            const guid = (this.playlist as Playlist | Album).guid;
+            this.songSource = this.type === 'playlist'
+                ? new PlaylistSongSource(this.playlistService, guid)
+                : new AlbumSongSource(this.albumService, guid);
+
+            this.fetchData(guid);
+        }
     }
 
+    /** Initialize SEO meta tags */
+    private initSeo(): void {
+        if (!this.playlist) return;
+
+        if (this.type === 'playlist') {
+            this.seoService.setPlaylistSeoData({
+                tracksCount: (this.playlist as Playlist).songCount,
+                creator: (this.playlist as Playlist).createdByName
+            });
+            this.seoService.setSeoData({
+                title: this.playlist.title,
+                description: `Listen to ${this.playlist.title} on TuneSync.`,
+                ogImage: { url: this.playlist.thumbnailUrl || '', alt: this.playlist.title },
+                ogType: 'music.playlist'
+            });
+        } else {
+            this.seoService.setAlbumSeoData({
+                tracksCount: (this.playlist as Album).songCount,
+                musician: (this.playlist as Album).artist?.guid || ''
+            });
+            this.seoService.setSeoData({
+                title: this.playlist.title,
+                description: `Listen to ${this.playlist.title} on TuneSync.`,
+                ogImage: { url: this.playlist.thumbnailUrl || '', alt: this.playlist.title },
+                ogType: 'music.album'
+            });
+        }
+    }
 
     fetchData(guid: string): void {
-
         if (!this.songSource) {
             this.notificationService.show('Song source is not initialized', 'error');
             return;
         }
-
-    
         this.songSource.loadInitial().subscribe();
-
-        const serviceDetailsCall = this.type === 'playlist'
-            ? this.playlistService.getPlaylistDetailsByGuid(guid)
-            : this.albumService.getAlbumDetailtByGuid(guid);
-
-        serviceDetailsCall.subscribe(resp => {
-            this.playlist = resp;
-            this.cdRef.detectChanges();
-        });
-
     }
 
     //TODO: Add infini-scroll. Example of infini-scroll can be found in All artists list
@@ -91,15 +111,9 @@ export class PlaylistComponent implements OnInit {
     }
 
     toggleFavorite(playlist: Playlist | null) {
-        if (playlist === null) {
-            return;
-        }
-
-        if (this.isAlbum(playlist)) {
-            this.albumService.toggleFavorite(playlist);
-        } else {
-            this.playlistService.toggleFavorite(playlist);
-        }
+        if (!playlist) return;
+        if (this.isAlbum(playlist)) this.albumService.toggleFavorite(playlist);
+        else this.playlistService.toggleFavorite(playlist);
     }
 
     isAlbum(playlist: Playlist | Album): playlist is Album {
@@ -107,35 +121,25 @@ export class PlaylistComponent implements OnInit {
     }
 
     openDeletePlaylistModal(modalContent: TemplateRef<any>) {
-        this.modalService.openModalFromTemplate(modalContent).then(
-            (result) => {
-                if (result === 'Yes') {
-                    this.deletePlaylist();
-                }
-            },
-            (reason) => {
-            }
-        );
+        this.modalService.openModalFromTemplate(modalContent).then(result => {
+            if (result === 'Yes') this.deletePlaylist();
+        });
     }
 
     deletePlaylist() {
-        if (this.playlist) {
-            let serviceCall = this.type === 'album'
-                ? this.albumService.deleteAlbum(this.playlist.guid)
-                : this.playlistService.deletePlaylist(this.playlist.guid);
+        if (!this.playlist) return;
 
-            serviceCall.subscribe({
-                next: (result) => {
-                    if (result) {
-                        this.router.navigate(["/"]);
-                        this.notificationService.show(`${this.type === 'album' ? "Album" : "Playlist"} deleted successfully!`, 'success');
-                    }
-                },
-                error: (err) => {
-                    this.notificationService.handleError(err);
-                }
-            });
-        }
+        const serviceCall = this.type === 'album'
+            ? this.albumService.deleteAlbum(this.playlist.guid)
+            : this.playlistService.deletePlaylist(this.playlist.guid);
+
+        serviceCall.subscribe({
+            next: () => {
+                this.router.navigate(['/']);
+                this.notificationService.show(`${this.type === 'album' ? 'Album' : 'Playlist'} deleted successfully!`, 'success');
+            },
+            error: (err) => this.notificationService.handleError(err)
+        });
     }
 
     addToMix(playlist: Playlist | Album) {
@@ -145,42 +149,38 @@ export class PlaylistComponent implements OnInit {
     }
 
     play() {
-        if (this.songSource) {
-            this.audioService.setCurrentSongSource(this.songSource);
-        }
+        if (this.songSource) this.audioService.setCurrentSongSource(this.songSource);
     }
 
     downloadMissingSongs(album: Album) {
-        console.log(album);
         const sourceId = extractAlbumId(album.sourceUrl);
-        if (sourceId) {
-            this.playlistService.downloadYoutubePlaylist(sourceId).then(jobId => {
-                if (!jobId) return;
-                this.isDownloading = true;
-                interval(5000)
-                    .pipe(
-                        switchMap(() => this.jobService.getJobStatus<string>(jobId)),
-                        takeWhile((jobResponse) => jobResponse.jobStatus !== 'Succeeded' && jobResponse.jobStatus !== 'Failed', true)
-                    )
-                    .subscribe({
-                        next: (jobResponse) => {
-                            if (jobResponse.jobStatus === 'Succeeded') {
-                                this.notificationService.show('Missing songs have been added successfully!', 'success');
-                                window.location.reload(); // TODO: REFRESH DATA WITHOUT RELOADING WHOLE PAGE
-                                this.isDownloading = false;
-                            } else if (jobResponse.jobStatus === 'Failed') {
-                                this.notificationService.show('Failed to download missing songs', 'error');
-                                this.isDownloading = false;
-                            }
-                        },
-                        error: (error) => {
+        if (!sourceId) return;
+
+        this.playlistService.downloadYoutubePlaylist(sourceId).then(jobId => {
+            if (!jobId) return;
+
+            this.isDownloading = true;
+            interval(5000)
+                .pipe(
+                    switchMap(() => this.jobService.getJobStatus<string>(jobId)),
+                    takeWhile(jobResponse => jobResponse.jobStatus !== 'Succeeded' && jobResponse.jobStatus !== 'Failed', true)
+                )
+                .subscribe({
+                    next: jobResponse => {
+                        if (jobResponse.jobStatus === 'Succeeded') {
+                            this.notificationService.show('Missing songs have been added successfully!', 'success');
+                            if (this.safeBrowserService.isBrowserPlatform) window.location.reload();
                             this.isDownloading = false;
-                            this.notificationService.handleError(error);
+                        } else if (jobResponse.jobStatus === 'Failed') {
+                            this.notificationService.show('Failed to download missing songs', 'error');
+                            this.isDownloading = false;
                         }
-                    });
-            });
-        }
+                    },
+                    error: error => {
+                        this.isDownloading = false;
+                        this.notificationService.handleError(error);
+                    }
+                });
+        });
     }
-
-
 }
